feat(web): make dice roller detachable into standalone window
